Use substr method to determine file extension
Alternatively use the "find_last_of" & "substr" methods on this
conveniently and performance wise.

This fixes the "." problem where we are working currently on a
directory that contains ".", which prior to this will get rid of any
occurences of this character from the original path (when using
std::accumulate without a binary operation).